Altered the GraphRDB code so that it threw the new shared exceptions AlreadyExists and DoesNotExist. This makes testing for error detection easier and avoids embedding explanatory text in the exception message.
Updated GraphFactory with two new methods, createGraph(String) and
openGraph(String), and invented GraphRDBFactory(IDBConnection) which
produces a graph factory where the graphs are persistent on the
database identified by the connection.

We have a strange test failure, so the graph Factory tests have been
switched off for the moment.