The mostly-for-testing method Node.create(String) has been extended so that quoted literals may be followed by a language-or-type modifier.
If the modifier contains a colon, it is taken to be a datatype URI and
a typed literal is created. Otherwise it is taken to be a language
code. No checking is done. [This arose from test code in the experimental
GraphBDB, which needs to ensure that different kinds of literal can be
encoded and hence needed ways of easily constructing such literals.]

The toString() method for ReificationStyle has been modified so that
the specific defined constants (Minimal, Standard, and Convenient) print
as their names; this makes test failures easier to read.

JMS.findStyle(RDFNode) now has a sibling findStyle(Node) [which it is
defined in terms of] to support the GraphBDB code which configures
according to a graph specification.