Rewrite now thinks that / and : are plain characters.
Added RuleReasonerFactory as an interface for ReasonerFactories than
can setRules() some rules to plant in the reasoners they create. Added
BaseRuleReasonerFactory as an implementation of that bit. Added
TestSetRules as a new test suite for this functionality. Had some of
the rule-based reasoner factories implement thus, adding addRules()
to the FBRuleReasoner to allow rules to be added on top of an existing
set. All this is to support ModelSpecs having ReasonerSpecs that can
include ruleset specifications.