Corrected bug 1163195 - we were not testing GraphRDBMaker because of a typo in the TestPackage file.
Subtlty: the `connection` argument was initialised in its declaration.
This introduced a timing error, since all the TestGraphRDBMaker connections
were created when the test suite was constructed. This was OK when running
the test alone, but failed when TestPackage was run, since the originally-
clean DB had by then got stuff left in it. Oops.

Fixed by initialising in setUp instead. (And *before* the call to
super.setUp, since that needs the connection to be open, because it
wants a GraphMaker.)